requirements.txt
openai

html2text
bs4
pyyaml
